Testing: refactor: organization of files
